Move modern http requests to own file
