+++ app/src/org/commcare/engine/references/JavaHttpRoot.java
@@ -10,7 +10,7 @@ import org.javarosa.core.reference.ReferenceFactory;
  */
 public class JavaHttpRoot implements ReferenceFactory {
 
-    private final HttpRequestGenerator generator = new HttpRequestGenerator();
+    private final HttpRequestGenerator generator = HttpRequestGenerator.buildNoAuthGenerator();
 
     @Override
     public Reference derive(String URI) throws InvalidReferenceException {
